fixed: migration, seeds, devise, manage product attributes

+++ db/migrate/20151014020627_remove_short_description_from_shoppe_product_translations.rb
@@ -1,5 +1,7 @@
 class RemoveShortDescriptionFromShoppeProductTranslations < ActiveRecord::Migration
   def change
-    remove_column :shoppe_product_translations, :short_description, :text
+    if column_exists? :shoppe_product_translations, :short_description
+      remove_column :shoppe_product_translations, :short_description, :text
+    end
   end
 end
